requirements.txt
aniso8601==9.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
distlib==0.3.4
docopt==0.6.2
easyocr==1.5.0
filelock==3.7.1
Flask==2.1.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
imageio==2.21.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
networkx==2.8.5
numpy==1.23.0
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
python-bidi==0.4.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
scikit-image==0.19.3
scipy==1.9.0
six==1.16.0
SQLAlchemy==1.4.39
tifffile==2022.8.8
torch==1.12.1
torchvision==0.13.1
typing-extensions==4.3.0
urllib3==1.26.11
virtualenv==20.15.1
Werkzeug==2.2.0
yarg==0.1.9
